Some mistakes

+++ ISatck.h
@@ -8,7 +8,7 @@ public:
 	virtual const T& top()const = 0;
 	virtual T pop() = 0;
 	virtual bool empty()const = 0;
-	virtual size_t size()const = 0;
+	virtual unsigned int size()const = 0;
 	virtual void clear() = 0;
 	virtual ~IStack() {};
 };
